20170518 Arrangemang add(post) och delete är klart!

+++ Controllers/CrudArrangemangController.vb
@@ -2,23 +2,31 @@
 Public Class CrudArrangemangController
     Private _dalobj As New arrangemangDAL
 
-    Public Function addArrangemang(userid As Integer, arrData As arrangemangInfo) As arrangemangcontainerInfo
 
+    Public Function addArrangemang(arrData As arrangemangInfo) As arrangemangcontainerInfo
+        Dim tmparrlist As New List(Of arrangemangInfo)
+        Dim nyttarr As New arrangemangInfo
         Dim ret As New arrangemangcontainerInfo
-        Try
-            arrData.Arrid = _dalobj.addArrangemang(userid, arrData)
-            If arrData.Arrid > 0 Then
+        Dim tmputovarid As Integer = 0
+        Dim contentid As Integer = 0
 
+        Try
+            arrData.Arrid = _dalobj.addArrangemang(arrData)
+            If arrData.Arrid > 0 Then
                 If arrData.Utovare <= 0 Then
-                    Dim tmputovarid As Integer = _dalobj.addutovare(arrData.UtovareData)
+                    tmputovarid = _dalobj.addutovare(arrData.UtovareData)
                     _dalobj.updateArrUtovare(arrData.Arrid, tmputovarid)
                 End If
 
-                Dim contentid As Integer = _dalobj.addArrangemangContent(arrData)
+                contentid = _dalobj.addArrangemangContent(arrData)
                 If contentid > 0 Then
                     If _dalobj.addContentToArr(arrData.Arrid, contentid) Then
-
                         If _dalobj.addFaktaToArrangemang(arrData) Then
+
+                            'skicka tillbaka dom nya värdena så som arrid till anropande funktion
+                            nyttarr.Arrid = arrData.Arrid
+                            tmparrlist.Add(nyttarr)
+                            ret.Arrangemanglist = tmparrlist
                             ret.Status = "OK! Arrangemanget Inlagt!"
                         Else
                             ret.Status = "Error fel vid inläggning av fakta!"
@@ -33,7 +41,6 @@ Public Class CrudArrangemangController
                 ret.Status = "Error fel vid inläggning av arrangemang!"
             End If
         Catch ex As Exception
-
             ret.Status = "Error fel! Arrangemanget har inte lagts in!!"
         End Try
 
@@ -41,16 +48,22 @@ Public Class CrudArrangemangController
 
     End Function
     Public Function delarrangemang(arrid As Integer) As Boolean
+        Dim ret As Boolean = False
+
         Try
-            If _dalobj.DeleteArrToContentID(arrid) Then
-                Return _dalobj.DeleteArrangemang(arrid)
-            Else
-                Return False
+            If _dalobj.DoArrangemangExist(arrid) Then
+                If _dalobj.DeleteArrToContentID(arrid) Then
+                    ret = _dalobj.DeleteArrangemang(arrid)
+                Else
+                    ret = False
+                End If
             End If
+
         Catch ex As Exception
-            Return False
+            ret = False
         End Try
 
+        Return ret
     End Function
 
 End Class
